Add shared preferences, but WIP
