+++ assets/json/timeTable.json
@@ -1,12 +1,7 @@
 [
   {
     "day": "27",
-    "time": "13:28",
-    "act1": "Britney Spears"
-  },
-  {
-    "day": "27",
-    "time": "13:30",
+    "time": "14:00",
     "act1": "Die Doofen"
   }
 ]
