+++ pubspec.yaml
@@ -40,19 +40,11 @@ dependencies:
   url_launcher: ^6.1.10 #https://pub.dev/packages/url_launcher
   expandable: ^5.0.1
   fluttertoast: ^8.2.1
-  device_calendar: ^4.3.1
-  html: ^0.15.2
-  shared_preferences: ^2.1.0
-  firebase_core: ^2.9.0
-  firebase_analytics: ^10.2.0
-  firebase_auth: ^4.4.1
-  cloud_firestore: ^4.5.1
-  rxdart: ^0.27.7
   flare_splash_screen: ^4.0.0
-  provider: ^6.0.5
   like_button: ^2.0.5
   flutter_local_notifications: ^12.0.4
   flutter_timezone: ^1.0.4
+  shared_preferences: ^2.1.0
 
 dev_dependencies:
   flutter_test:
